Fleshed out Operations section.
This now has added a couple new data structures to deal with operations and the debounch thereof.  Basically, when we detect a op pin pulled high we set it's byte to DELAY_OP, we only fire ops that are set to DELAY_OP and we only check ops when byte has been decremented back to zero, this allows independent command debounce for each command, for now all commands have the same debounce value but could even have separate values there as well.